services/catalog-service/requirements.txt
fastapi==0.115.4
setuptools>=75.0.0,<81
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-httpx==0.48b0
python-json-logger==2.0.7
